requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
contextlib2==21.6.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
MedPy==0.5.2
ml_collections==0.1.1
monai==1.5.1
mpmath==1.3.0
narwhals==2.14.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==2.12.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
simpleitk==2.5.3
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
stringzilla==4.5.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2024.8.30
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.4
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.4
zipp==3.23.0
